% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_to_celsius.R
\name{conv_to_celsius}
\alias{conv_to_celsius}
\title{Convert temperture captured in Fahrenheit to celsius}
\usage{
conv_to_celsius(.x)
}
\arguments{
\item{.x}{the field containing temperatures to convert}
}
\value{
A data frame
}
\description{
Convert temperture captured in Fahrenheit to celsius
}
