% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_type_ck.R
\name{dp_type_ck}
\alias{dp_type_ck}
\title{Title Check data types during processing to insure consistency}
\usage{
dp_type_ck(.data, .table_type)
}
\arguments{
\item{.data}{A data frame who's name contains either "site", "fish", or "water"}

\item{.table_type}{The type of table to be checked either "site", "fish", or "water"}
}
\value{
A message or error
}
\description{
Title Check data types during processing to insure consistency
}
