% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_other_na.R
\name{field_other_na}
\alias{field_other_na}
\title{Check for NA values in important but not required fields}
\usage{
field_other_na(.data, .table)
}
\arguments{
\item{.data}{A data frame containing either site data or fish data}

\item{.table}{A string designating the table type ("site" or "fish")}
}
\value{
A message or warning and table
}
\description{
Check for NA values in important but not required fields
}
