% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_rqd_na.R
\name{field_rqd_na}
\alias{field_rqd_na}
\title{Check for NA values in required fields}
\usage{
field_rqd_na(.data, .table)
}
\arguments{
\item{.data}{A data frame containing either site data or fish data}

\item{.table}{A string designating the table type ("site" or "fish")}
}
\value{
A message or warning and table
}
\description{
Check for NA values in required fields
}
