% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_xlsx_wkbk.R, R/field_xlsx_wkbk_v2.R
\name{field_xlsx_wkbk}
\alias{field_xlsx_wkbk}
\title{Save transformed field data to the STReaMS upload template}
\usage{
field_xlsx_wkbk(
  .site_data,
  .rare_data = NULL,
  .rare_data_2 = NULL,
  .ntf_data = NULL,
  .year,
  .study,
  .agency,
  .output_path,
  .overwrite
)

field_xlsx_wkbk(
  .site_data,
  .rare_data = NULL,
  .rare_data_2 = NULL,
  .ntf_data = NULL,
  .year,
  .study,
  .agency,
  .output_path,
  .overwrite
)
}
\arguments{
\item{.site_data}{A data frame containing spec'd site data}

\item{.rare_data}{A data frame containing spec'd rare fish data}

\item{.rare_data_2}{A data frame containing spec'd rare fish data overflow records}

\item{.ntf_data}{A data frame containing spec'd nontagged fish data}

\item{.year}{Year data was collected (numeric)}

\item{.study}{The study underwhich the data was collected}

\item{.agency}{The agency managing the study}

\item{.output_path}{Path to data destination directory}

\item{.overwrite}{Overwrite existing data if it exists?}
}
\value{
Nothing, creates an xlsx file in destination directory

Nothing, creates an xlsx file in destination directory
}
\description{
Save transformed field data to the STReaMS upload template

Save transformed field data to the STReaMS upload template
}
