% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_config.R
\name{get_config}
\alias{get_config}
\title{Fetch the configuration file for a specific project}
\usage{
get_config(.config_name, .machine = "FWS", .study)
}
\arguments{
\item{.config_name}{Name of the target config file}

\item{.machine}{The machine you are working on (FWS or CNHP)}

\item{.study}{The target study}
}
\value{
A list
}
\description{
Fetch the configuration file for a specific project
}
