% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orphan.R
\name{orphan}
\alias{orphan}
\title{Title Check hierarchical data for orphaned records}
\usage{
orphan(.parent, .child, .key, .write2csv = FALSE)
}
\arguments{
\item{.parent}{The parent table}

\item{.child}{The child table}

\item{.key}{The primary key on the parent tabled referenced as a foreign key
on the child table}

\item{.write2csv}{Write orphaned records to root/orphan.csv?}
}
\value{
A warning and .csv (if requested) or a success message
}
\description{
Title Check hierarchical data for orphaned records
}
\examples{
parent <- data.frame(key = 1:5,
                    varA = letters[1:5])
child <- data.frame(key = c(1,2,5,7,10),
                   varB = LETTERS[6:10])
orphan(.parent = parent, .child = child, .key = "key")
}
