% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_datetime.R
\name{parse_datetime}
\alias{parse_datetime}
\title{Parse datetimes from incorrectly formatted values}
\usage{
parse_datetime(.x)
}
\arguments{
\item{.x}{A datetime stored as a character string}
}
\value{
A POSIXct datetime
}
\description{
Parse datetimes from incorrectly formatted values
}
