% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_streams_stock_sub.R
\name{read_streams_stock_sub}
\alias{read_streams_stock_sub}
\title{Import a STReaMS stocking data submission}
\usage{
read_streams_stock_sub(.file_path)
}
\arguments{
\item{.file_path}{A string or vector containing path to target datasets}
}
\value{
A list
}
\description{
Import a STReaMS stocking data submission
}
