% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_ck_ranges.R
\name{stock_ck_ranges}
\alias{stock_ck_ranges}
\title{Test ranges for important numeric fields on stocking submissions}
\usage{
stock_ck_ranges(.data)
}
\arguments{
\item{.data}{A stocking submission dataset}
}
\value{
Test results
}
\description{
Test ranges for important numeric fields on stocking submissions
}
