% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_ck_tags.R
\name{stock_ck_tags}
\alias{stock_ck_tags}
\title{Check for duplicated or missing values in pittag column of the stocking submission dataset}
\usage{
stock_ck_tags(.data, .tag_col)
}
\arguments{
\item{.data}{A data frame}

\item{.tag_col}{The field containing pittag codes}
}
\value{
A message or warning
}
\description{
Check for duplicated or missing values in pittag column of the stocking submission dataset
}
