% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_flag.R
\name{stock_flag}
\alias{stock_flag}
\title{Apply QC flags to stocking dataset}
\usage{
stock_flag(.data)
}
\arguments{
\item{.data}{A stocking dataset formatted to STReaMS spec}
}
\value{
A dataframe and message or warning
}
\description{
Apply QC flags to stocking dataset
}
