% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_read_streams_sub.R
\name{stock_read_streams_sub}
\alias{stock_read_streams_sub}
\title{Import a STReaMS stocking data submission}
\usage{
stock_read_streams_sub(.file_path)
}
\arguments{
\item{.file_path}{A string or vector containing path to target datasets}
}
\value{
A list
}
\description{
Import a STReaMS stocking data submission
}
