% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_xlsx_wkbk.R
\name{stock_xlsx_wkbk}
\alias{stock_xlsx_wkbk}
\title{Create data typed STReaMS xlsx submission files}
\usage{
stock_xlsx_wkbk(
  .stock_data,
  .event_data,
  .shed_data = NULL,
  .mort_data = NULL,
  .output_path,
  .overwrite
)
}
\arguments{
\item{.stock_data}{A data frame containing QC'd stocking data}

\item{.event_data}{A data frame containing QC'd stocking_event data}

\item{.shed_data}{A data frame containing individuals who shed their tags}

\item{.mort_data}{A data frame containing any dead individuals}

\item{.output_path}{A file path to save the xlsx}

\item{.overwrite}{Do you want to overwrite an existing .xlsx (T/F)}
}
\value{
An xlsx workbook
}
\description{
Create data typed STReaMS xlsx submission files
}
