% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_parse_enc.R
\name{str_parse_enc}
\alias{str_parse_enc}
\title{Read in and parse STReaMS encounters download files}
\usage{
str_parse_enc(.path)
}
\arguments{
\item{.path}{A string representing the file path to the dataset}
}
\value{
A data frame
}
\description{
Read in and parse STReaMS encounters download files
}
