% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STReaMS_rqd_na.R
\name{STReaMS_rqd_na}
\alias{STReaMS_rqd_na}
\title{Check STReaMS submissions for missing values in required fields}
\usage{
STReaMS_rqd_na(.data, .table)
}
\arguments{
\item{.data}{An imported STReaMS data submission}

\item{.table}{The type of table ("site", "rare" or "ntf")}
}
\value{
a table and message
}
\description{
Check STReaMS submissions for missing values in required fields
}
