% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_aggregate.R
\name{tag_aggregate}
\alias{tag_aggregate}
\title{Title Aggregate ppit tage detections to specified time interval}
\usage{
tag_aggregate(.data, .interval = "min")
}
\arguments{
\item{.data}{A parsed pit tag file}

\item{.interval}{Time interval to aggregate to}
}
\value{
A data frame
}
\description{
Title Aggregate ppit tage detections to specified time interval
}
\examples{
tag_aggregate(tbl_ex_parsed_xlsx, .interval = "hour")
tag_aggregate(tbl_ex_parsed_dat, .interval = "day")
}
