% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_calc_effort.R
\name{tag_calc_effort}
\alias{tag_calc_effort}
\title{Title Calculate the effort in hours a portable submersible (PS) antenna was
operational}
\usage{
tag_calc_effort(.data, .tag_prefix = "^3E7")
}
\arguments{
\item{.data}{A PS file downloaded using Biomark TagManager and parsed with
UCRBDataTools:: tag_parse_xlsx or tag_parse_txt}

\item{.tag_prefix}{The "test tag" prefix}
}
\value{
A data frame
}
\description{
Title Calculate the effort in hours a portable submersible (PS) antenna was
operational
}
\examples{
tag_calc_effort(.data = tbl_ex_parsed_xlsx,
                .tag_prefix = "3E7")
}
