% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_parse_dat.R
\name{tag_parse_dat}
\alias{tag_parse_dat}
\title{Title Parse Perminant PIA download (.dat) files}
\usage{
tag_parse_dat(.path)
}
\arguments{
\item{.path}{File path including file name}
}
\value{
A data frame
}
\description{
Title Parse Perminant PIA download (.dat) files
}
\examples{
tag_parse_dat(.path = system.file("extdata",
                                  "GreenRiverTusherDiversion_Tags.dat",
                                  package = "UCRBDataTools"))
}
