% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_parse_xlsx.R
\name{tag_parse_xlsx}
\alias{tag_parse_xlsx}
\title{Read and parse raw tag manager download file in .xlsx format}
\usage{
tag_parse_xlsx(.path, .sheet = 1)
}
\arguments{
\item{.path}{Path to and file name of target excel file}

\item{.sheet}{Which sheet to read}
}
\value{
A tibble or data.frame
}
\description{
Read and parse raw tag manager download file in .xlsx format
}
\examples{
tag_parse_xlsx(system.file("extdata",
                           "2022_CASTLE_BUTTE_PIA_1.xlsx",
                           package = "UCRBDataTools"))
}
