% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_rm_test_tag.R
\name{tag_rm_test_tags}
\alias{tag_rm_test_tags}
\title{Title Remove test-tags}
\usage{
tag_rm_test_tags(.data, .tag_prefix = "3E7", .tag_col = cd_hex)
}
\arguments{
\item{.data}{A data.frame or tibble}

\item{.tag_prefix}{The first 3 characters of the hex string}

\item{.tag_col}{The column containing the hex strings}
}
\value{
A tibble or data.frame
}
\description{
Title Remove test-tags
}
\examples{
tag_rm_test_tags(.data = tbl_ex_parsed_dat,
                .tag_prefix = "3E7",
                .tag_col = cd_hex)
tag_rm_test_tags(.data = tbl_ex_parsed_xlsx,
                .tag_prefix = "3E7",
                .tag_col = cd_hex)
}
