% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagmgr_parse_xlsx.R
\name{tagmgr_parse_xlsx}
\alias{tagmgr_parse_xlsx}
\title{Read and parse raw tag manager download file}
\usage{
tagmgr_parse_xlsx(.path, .sheet = 1)
}
\arguments{
\item{.path}{Path to and file name of target excel file}

\item{.sheet}{Which sheet to read}
}
\value{

}
\description{
Read and parse raw tag manager download file
}
\examples{
tagmgr_parse_xlsx(system.file("extdata",
                              "2022_Lake_Powell_PS_Across_from_Ticcaboo.xlsx",
                              package = "UCRBDataTools"))
}
