% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tbl_ex_parsed_dat}
\alias{tbl_ex_parsed_dat}
\title{Example of parsed .dat file}
\format{
A table with 13,486 rows and 7 variables
}
\usage{
tbl_ex_parsed_dat
}
\description{
A dataset containing the contents of a parsed Perminant antenna .dat file
}
\keyword{datasets}
