% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tbl_ex_parsed_xlsx}
\alias{tbl_ex_parsed_xlsx}
\title{Example of a parsed .xlsx file}
\format{
A table with 890 rows and 7 variables
}
\usage{
tbl_ex_parsed_xlsx
}
\description{
A dataset containing the contents of a parsed Biomark TagManager download file
exported to .xlsx
}
\keyword{datasets}
