% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tbl_NWIS_gage}
\alias{tbl_NWIS_gage}
\title{National Water Information System gage locatons and codes for STReaMS rivers}
\format{
A data frame
}
\usage{
tbl_NWIS_gage
}
\description{
A dataset containing ~438 records of location and other high level station data
}
\keyword{datasets}
