% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tbl_NWIS_guage}
\alias{tbl_NWIS_guage}
\title{Local USGS guaging stations}
\format{
A data frame
}
\usage{
tbl_NWIS_guage
}
\description{
A dataset containing 13 records of location and other high level station data
}
\keyword{datasets}
