% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tbl_NWIS_params}
\alias{tbl_NWIS_params}
\title{National Water Information System data parameter codes}
\format{
A data frame
}
\usage{
tbl_NWIS_params
}
\description{
A dataset 11 useful data parameter codes for use with dataRetrevial
}
\keyword{datasets}
