% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tbl_rvr_crswlk}
\alias{tbl_rvr_crswlk}
\title{River codes, names and limited GNIS ID's adapted from STReaMS species look-up table}
\format{
A data frame
}
\usage{
tbl_rvr_crswlk
}
\description{
A dataset with 95 records and 4 fields
}
\keyword{datasets}
