% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tbl_spp}
\alias{tbl_spp}
\title{Species definitions adapted from STReaMS species look-up table}
\format{
A data frame
}
\usage{
tbl_spp
}
\description{
A dataset with 83 records and 5 fields
}
\keyword{datasets}
