% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tbl_spp_crswlk}
\alias{tbl_spp_crswlk}
\title{Species definitions adapted from STReaMS species look-up table matched with
CPW species codes}
\format{
A data frame
}
\usage{
tbl_spp_crswlk
}
\description{
A dataset with 79 records and 6 fields
}
\keyword{datasets}
