% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tmplt_adamas_field}
\alias{tmplt_adamas_field}
\title{A series of dataframes useful in generating ADAMAS data submissions}
\format{
A list
}
\usage{
tmplt_adamas_field
}
\description{
A series of dataframes useful in generating ADAMAS data submissions
}
\keyword{datasets}
