% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tmplt_dp_standard}
\alias{tmplt_dp_standard}
\title{Template to check processed data's data types}
\format{
A list containing 3 typed dataframes
}
\usage{
tmplt_dp_standard
}
\description{
A dataset containing 1 record (NA) but possessing typed columns.  This may be
joined to processed data to enusre no unexpected changes to data types
}
\keyword{datasets}
