% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tmplt_electrofish_standard}
\alias{tmplt_electrofish_standard}
\title{Template to check data against the electrofishing data standard}
\format{
A list containing 3 typed dataframes
}
\usage{
tmplt_electrofish_standard
}
\description{
A dataset containing 1 record (NA) but possessing typed columns.  This may be
joined to processed data to ensure no unexpected changes to data types
This template superceeds \code{tmplt_dp_standard}
}
\keyword{datasets}
