% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tmplt_electrofish_standard_db_v2}
\alias{tmplt_electrofish_standard_db_v2}
\title{Template to check data against the electrofishing database schema}
\format{
A list containing 3 typed dataframes
}
\usage{
tmplt_electrofish_standard_db_v2
}
\description{
A dataset containing 1 record (NA) but possessing typed columns.  This may be
joined to processed data to ensure no unexpected changes to data types
}
\keyword{datasets}
