% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tmplt_electrofish_standard_v2}
\alias{tmplt_electrofish_standard_v2}
\title{Template to check data against the electrofishing data standard}
\format{
A list containing 3 typed dataframes and 3 column type strings
}
\usage{
tmplt_electrofish_standard_v2
}
\description{
A dataset containing 1 record (NA) but possessing typed columns.  This may be
joined to processed data to ensure no unexpected changes to data types
Also contains column types strings for import with readr::read_csv
This template superceeds \code{tmplt_electrofish_standard}
}
\keyword{datasets}
