% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tmplt_heritage}
\alias{tmplt_heritage}
\title{A datatyped dataframe for heritage data submission generation}
\format{
An empty dataframe
}
\usage{
tmplt_heritage
}
\description{
A datatyped dataframe for heritage data submission generation
}
\keyword{datasets}
