% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tmplt_streams_field}
\alias{tmplt_streams_field}
\title{Data typed streams field data upload template, including vectors of rqd fields
and more}
\format{
A list
}
\usage{
tmplt_streams_field
}
\description{
A list, 9 elements
}
\keyword{datasets}
