% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dimension_streams}
\alias{dimension_streams}
\title{Data typed adamas field data upload template}
\format{
A list
}
\usage{
dimension_streams
}
\description{
A list, 3 elements (dataframes)
}
\keyword{datasets}
