% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dimension_streams}
\alias{dimension_streams}
\title{A list including relevant dimension tables and vectors from the STReaMS database}
\format{
A list
}
\usage{
dimension_streams
}
\description{
A large object containing 19 elements
}
\keyword{datasets}
