% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tmplt_adamas_field}
\alias{tmplt_adamas_field}
\title{A list including relevant dimension tables and vectors from the STReaMS database}
\format{
A list
}
\usage{
tmplt_adamas_field
}
\description{
A large object containing 19 elements
}
\keyword{datasets}
