% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tmplt_electrofish_standard}
\alias{tmplt_electrofish_standard}
\title{Template to check data against the STReaMS stocking template data standard}
\format{
A list containing 2 typed dataframes and a series of 4 vectors
referencing column names
}
\usage{
tmplt_electrofish_standard
}
\description{
Template to check data against the STReaMS stocking template data standard
}
\keyword{datasets}
